+++ backend/src/main/java/com/hiddenartist/backend/domain/genre/persistence/MentorGenre.java
@@ -1,6 +1,6 @@
 package com.hiddenartist.backend.domain.genre.persistence;
 
-import com.hiddenartist.backend.domain.mentoring.persistence.Mentor;
+import com.hiddenartist.backend.domain.mentor.persistence.Mentor;
 import com.hiddenartist.backend.global.type.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
