+++ backend/src/main/java/com/hiddenartist/backend/domain/mentor/persistence/type/CertificationStatus.java
@@ -1,0 +1,5 @@
+package com.hiddenartist.backend.domain.mentor.persistence.type;
+
+public enum CertificationStatus {
+  UNVERIFIED, VERIFIED
+}
