+++ backend/src/main/java/com/hiddenartist/backend/domain/mentoring/controller/response/MentorResponse.java
@@ -2,7 +2,7 @@ package com.hiddenartist.backend.domain.mentoring.controller.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.hiddenartist.backend.domain.account.persistence.Account;
-import com.hiddenartist.backend.domain.mentoring.persistence.Mentor;
+import com.hiddenartist.backend.domain.mentor.persistence.Mentor;
 
 public record MentorResponse(
     String name,
