+++ backend/src/main/java/com/hiddenartist/backend/domain/mentoring/persistence/MentoringApplication.java
@@ -1,5 +1,6 @@
 package com.hiddenartist.backend.domain.mentoring.persistence;
 
+import com.hiddenartist.backend.domain.account.persistence.Account;
 import com.hiddenartist.backend.domain.mentoring.persistence.type.MentoringApplicationStatus;
 import com.hiddenartist.backend.domain.payment.persistence.Payment;
 import com.hiddenartist.backend.global.converter.MentoringApplicationStatusConverter;
@@ -7,6 +8,7 @@ import com.hiddenartist.backend.global.type.BaseEntity;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToOne;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
@@ -23,8 +25,11 @@ public class MentoringApplication extends BaseEntity {
   @Convert(converter = MentoringApplicationStatusConverter.class)
   private MentoringApplicationStatus mentoringApplicationStatus;
 
-  @OneToOne(fetch = FetchType.LAZY)
+  @ManyToOne(fetch = FetchType.LAZY)
   private Mentoring mentoring;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  private Account account;
 
   @OneToOne(fetch = FetchType.LAZY)
   private Payment payment;
