--- backend/src/main/java/com/hiddenartist/backend/domain/mentoring/persistence/type/CertificationStatus.java
@@ -1,5 +1,0 @@
-package com.hiddenartist.backend.domain.mentoring.persistence.type;
-
-public enum CertificationStatus {
-  UNVERIFIED, VERIFIED
-}
