+++ backend/src/main/java/com/hiddenartist/backend/domain/search/controller/response/MentoringSearchResponse.java
@@ -1,8 +1,8 @@
 package com.hiddenartist.backend.domain.search.controller.response;
 
+import com.hiddenartist.backend.domain.mentor.persistence.Mentor;
 import com.hiddenartist.backend.domain.mentoring.controller.response.MentorResponse;
 import com.hiddenartist.backend.domain.mentoring.controller.response.MentoringResponse;
-import com.hiddenartist.backend.domain.mentoring.persistence.Mentor;
 import com.hiddenartist.backend.domain.mentoring.persistence.Mentoring;
 import lombok.Getter;
