+++ backend/src/main/java/com/hiddenartist/backend/global/converter/CertificationStatusConverter.java
@@ -1,6 +1,6 @@
 package com.hiddenartist.backend.global.converter;
 
-import com.hiddenartist.backend.domain.mentoring.persistence.type.CertificationStatus;
+import com.hiddenartist.backend.domain.mentor.persistence.type.CertificationStatus;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
